pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-control-center"
version = "0.1.0"
description = "Autonomous Agent Control Center - LangGraph-powered multi-agent orchestration"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.3.0",
    "langchain-community>=0.3.0",
    "langgraph>=0.2.0",
    "streamlit>=1.40.0",
    "tavily-python>=0.5.0",
    "httpx>=0.27.0",
    "mcp>=1.0.0",
    "networkx>=3.2",
    "matplotlib>=3.8.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
addopts = "-v"
testpaths = ["tests"]
asyncio_mode = "auto"
